Fix slowness with repetitive stars plus extras
also:
- allow for escaped characters
- use long for string lengths

+++ match.h
@@ -7,6 +7,6 @@
 
 #include <stdbool.h>
 
-bool match(const char *pat, int plen, const char *str, int slen);
+bool match(const char *pat, long plen, const char *str, long slen);
 
 #endif
